// src/consts/consts.d.ts
export const COMPANY: {
  SITE_TITLE: {
    full: string, small: string,
  },
  LEGAL_FULL_NAME: string,
  ITN: string,
  YEARS: string,
  MASTER: {
    NAME: string,
    SITE: string,
  },
};
export const LOCATION: string;
export const MODE_AUTH_COOKIE: boolean;
export const ENV: string;
export const GET_FROM_GOOGLE: boolean;